Separate libtest module
